Changed package name and updated cors config
