--- Backend/collabwithkatz/src/main/java/com/katziio/codewithkatz/controller/editorcontroller/EditorSocialMediaController.java
@@ -1,4 +1,0 @@
-package com.katziio.codewithkatz.controller.editorcontroller;
-
-public class EditorSocialMediaController {
-}
