--- Backend/collabwithkatz/src/main/java/com/katziio/codewithkatz/entity/editorentity/EditorPreferences.java
@@ -1,18 +1,0 @@
-package com.katziio.codewithkatz.entity.editorentity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "editor_preferences")
-public class EditorPreferences {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-    private String projectPreferences;
-}
