--- Backend/collabwithkatz/src/main/java/com/katziio/codewithkatz/entity/editorentity/EditorProject.java
@@ -1,19 +1,0 @@
-package com.katziio.codewithkatz.entity.editorentity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "editor_project")
-public class EditorProject {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-    private String projectName;
-    private boolean projectStatus;
-}
