--- Backend/collabwithkatz/src/main/java/com/katziio/codewithkatz/entity/editorentity/EditorSkill.java
@@ -1,22 +1,0 @@
-package com.katziio.codewithkatz.entity.editorentity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "editor_skill")
-public class EditorSkill {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-    private String aboutSkill;
-    private String availability;
-    private String flexibility;
-    private String technicalProficiency;
-    private String equipmentDetails;
-}
