+++ Backend/collabwithkatz/src/main/java/com/katziio/collabwithkatz/CollabWithKatzApplication.java
@@ -1,0 +1,15 @@
+package com.katziio.collabwithkatz;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+import java.io.FileNotFoundException;
+
+@SpringBootApplication
+public class CollabWithKatzApplication {
+
+    public static void main(String[] args)  {
+        SpringApplication.run(CollabWithKatzApplication.class, args);
+    }
+}
+
