+++ Backend/collabwithkatz/src/main/java/com/katziio/collabwithkatz/controller/editorcontroller/EditorSkillsController.java
@@ -1,0 +1,4 @@
+package com.katziio.collabwithkatz.controller.editorcontroller;
+
+public class EditorSkillsController {
+}
