+++ Backend/collabwithkatz/src/main/java/com/katziio/collabwithkatz/dto/editordto/EditorCommunicationLanguageDTO.java
@@ -1,0 +1,21 @@
+package com.katziio.collabwithkatz.dto.editordto;
+
+import com.katziio.collabwithkatz.entity.editorentity.EditorCommunicationLanguage;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class EditorCommunicationLanguageDTO {
+    private Long id;
+    private String language;
+    private int languageProficiency;
+    public EditorCommunicationLanguageDTO(EditorCommunicationLanguage editorCommunicationLanguage)
+    {
+        this.id=editorCommunicationLanguage.getId();
+        this.languageProficiency=editorCommunicationLanguage.getLanguageProficiency();
+        this.language=editorCommunicationLanguage.getLanguage();
+    }
+}
