+++ Backend/collabwithkatz/src/main/java/com/katziio/collabwithkatz/dto/editordto/EditorSampleVideoDTO.java
@@ -1,0 +1,22 @@
+package com.katziio.collabwithkatz.dto.editordto;
+
+import com.katziio.collabwithkatz.entity.editorentity.EditorSampleVideo;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class EditorSampleVideoDTO {
+    private Long id;
+    private String videoName;
+    private String videoLink;
+
+    public EditorSampleVideoDTO(EditorSampleVideo editorSampleVideo)
+    {
+        this.id=editorSampleVideo.getId();
+        this.videoLink=editorSampleVideo.getVideoLink();
+        this.videoName=editorSampleVideo.getVideoName();
+    }
+}
