+++ Backend/collabwithkatz/src/main/java/com/katziio/collabwithkatz/entity/editorentity/EditorPreferences.java
@@ -1,0 +1,18 @@
+package com.katziio.collabwithkatz.entity.editorentity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "editor_preferences")
+public class EditorPreferences {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+    private String projectPreferences;
+}
