+++ Backend/collabwithkatz/src/main/java/com/katziio/collabwithkatz/entity/editorentity/EditorSampleVideo.java
@@ -1,0 +1,19 @@
+package com.katziio.collabwithkatz.entity.editorentity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "editor_sample_video")
+public class EditorSampleVideo {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+    private String videoName;
+    private String videoLink;
+}
