+++ Backend/collabwithkatz/src/main/java/com/katziio/collabwithkatz/entity/editorentity/EditorSocialMedia.java
@@ -1,0 +1,19 @@
+package com.katziio.collabwithkatz.entity.editorentity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "editor_social_media")
+public class EditorSocialMedia {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+    private String platform;
+    private String platformLink;
+}
