--- Backend/collabwithkatz/src/test/java/com/katziio/codewithkatz/CollabWithKatzApplicationTests.java
@@ -1,13 +1,0 @@
-package com.katziio.codewithkatz;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CollabWithKatzApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
