+++ Backend/collabwithkatz/src/test/java/com/katziio/collabwithkatz/CollabWithKatzApplicationTests.java
@@ -1,0 +1,13 @@
+package com.katziio.collabwithkatz;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CollabWithKatzApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
